refactor: [#1541] rename mod
