+++ packages/http-tracker-core/benches/helpers/util.rs
@@ -15,7 +15,7 @@ use bittorrent_tracker_core::authentication::key::repository::in_memory::InMemor
 use bittorrent_tracker_core::authentication::service::AuthenticationService;
 use bittorrent_tracker_core::databases::setup::initialize_database;
 use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
-use bittorrent_tracker_core::torrent::repository::persisted::DatabaseDownloadsMetricRepository;
+use bittorrent_tracker_core::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
 use bittorrent_tracker_core::whitelist::authorization::WhitelistAuthorization;
 use bittorrent_tracker_core::whitelist::repository::in_memory::InMemoryWhitelist;
 use futures::future::BoxFuture;
