+++ packages/tracker-core/src/statistics/event/handler.rs
@@ -7,7 +7,7 @@ use torrust_tracker_torrent_repository::event::Event;
 
 use crate::statistics::repository::Repository;
 use crate::statistics::TRACKER_CORE_PERSISTENT_TORRENTS_DOWNLOADS_TOTAL;
-use crate::torrent::repository::persisted::DatabaseDownloadsMetricRepository;
+use crate::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
 
 pub async fn handle_event(
     event: Event,
