+++ packages/tracker-core/src/test_helpers.rs
@@ -20,7 +20,7 @@ pub(crate) mod tests {
     use crate::databases::setup::initialize_database;
     use crate::scrape_handler::ScrapeHandler;
     use crate::torrent::repository::in_memory::InMemoryTorrentRepository;
-    use crate::torrent::repository::persisted::DatabaseDownloadsMetricRepository;
+    use crate::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
     use crate::whitelist::repository::in_memory::InMemoryWhitelist;
     use crate::whitelist::{self};
