+++ packages/tracker-core/src/torrent/repository/mod.rs
@@ -1,3 +1,3 @@
 //! Torrent repository implementations.
 pub mod in_memory;
-pub mod persisted;
+pub mod downloads;
